+++ tests/compile_fail/unsize_to_different_lifetime.rs
@@ -1,0 +1,7 @@
+use bump_scope::{unsize_bump_box, BumpBox};
+
+fn evil_unsize(boxed: BumpBox<[i32; 3]>) -> BumpBox<'static, [i32]> {
+    unsize_bump_box!(boxed)
+}
+
+fn main() {}